Rechartsに渡すデータ周りの型を定義・更新
