--- src/types/index.ts
@@ -1,9 +1,0 @@
-export interface Prefecture {
-  prefCode: number;
-  prefName: string;
-}
-
-export interface PrefectureResponse {
-  message: null;
-  result: Prefecture[];
-}
